+++ src/store/index.js
@@ -1,9 +1,11 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import app from './modules/app'
+import emojiPacks from './modules/emojiPacks'
 import errorLog from './modules/errorLog'
-import moderationLog from './modules/moderationLog'
+import getters from './getters'
 import invites from './modules/invites'
+import moderationLog from './modules/moderationLog'
 import peers from './modules/peers'
 import permission from './modules/permission'
 import relays from './modules/relays'
@@ -14,8 +16,6 @@ import tagsView from './modules/tagsView'
 import user from './modules/user'
 import userProfile from './modules/userProfile'
 import users from './modules/users'
-import getters from './getters'
-import emojiPacks from './modules/emojiPacks.js'
 
 Vue.use(Vuex)
 
@@ -23,6 +23,7 @@ const store = new Vuex.Store({
   modules: {
     app,
     errorLog,
+    emojiPacks,
     moderationLog,
     invites,
     peers,
@@ -34,8 +35,7 @@ const store = new Vuex.Store({
     tagsView,
     user,
     userProfile,
-    users,
-    emojiPacks
+    users
   },
   getters
 })