+++ src/store/modules/emojiPacks.js
@@ -4,6 +4,7 @@ import {
   deleteEmojiFile,
   deletePack,
   downloadFrom,
+  fetchPack,
   importFromFS,
   listPacks,
   listRemotePacks,
@@ -16,19 +17,40 @@ import { Message } from 'element-ui'
 
 import Vue from 'vue'
 
-const packs = {
+const emojiPacks = {
   state: {
-    activeCollapseItems: [],
+    activeTab: '',
+    currentFilesPage: 1,
+    currentPage: 1,
+    filesPageSize: 30,
+    localPackFilesCount: 0,
     localPacks: {},
+    localPacksCount: 0,
+    pageSize: 50,
     remoteInstance: '',
     remotePacks: {}
   },
   mutations: {
-    SET_ACTIVE_COLLAPSE_ITEMS: (state, items) => {
-      state.activeCollapseItems = items
+    SET_ACTIVE_TAB: (state, tab) => {
+      state.activeTab = tab
+    },
+    SET_FILES_COUNT: (state, count) => {
+      state.localPackFilesCount = count
+    },
+    SET_FILES_PAGE: (state, page) => {
+      state.currentFilesPage = page
     },
     SET_LOCAL_PACKS: (state, packs) => {
       state.localPacks = packs
+    },
+    SET_LOCAL_PACKS_COUNT: (state, count) => {
+      state.localPacksCount = count
+    },
+    SET_PACK_FILES: (state, { name, files }) => {
+      state.localPacks = { ...state.localPacks, [name]: { ...state.localPacks[name], files }}
+    },
+    SET_PAGE: (state, page) => {
+      state.currentPage = page
     },
     SET_REMOTE_INSTANCE: (state, name) => {
       state.remoteInstance = name
@@ -67,10 +89,12 @@ const packs = {
 
       commit('UPDATE_LOCAL_PACK_FILES', { name: packName, files: result.data })
     },
-    async DeleteEmojiFile({ commit, getters }, { packName, shortcode }) {
-      let result
+    async DeleteEmojiFile({ commit, dispatch, getters, state }, { packName, shortcode }) {
+      const { [shortcode]: value, ...updatedPackFiles } = state.localPacks[packName].files
+      commit('UPDATE_LOCAL_PACK_FILES', { name: packName, files: updatedPackFiles })
+
       try {
-        result = await deleteEmojiFile(packName, shortcode, getters.authHost, getters.token)
+        await deleteEmojiFile(packName, shortcode, getters.authHost, getters.token)
       } catch (_e) {
         return
       }
@@ -79,8 +103,11 @@ const packs = {
         type: 'success',
         duration: 5 * 1000
       })
-
-      commit('UPDATE_LOCAL_PACK_FILES', { name: packName, files: result.data })
+      if (Object.keys(updatedPackFiles).length === 0 && state.currentFilesPage > 1) {
+        dispatch('FetchSinglePack', { name: packName, page: state.currentFilesPage - 1 })
+      } else {
+        dispatch('FetchSinglePack', { name: packName, page: state.currentFilesPage })
+      }
     },
     async CreatePack({ getters }, { name }) {
       await createPack(getters.authHost, getters.token, name)
@@ -89,7 +116,7 @@ const packs = {
       await deletePack(getters.authHost, getters.token, name)
     },
     async DownloadFrom({ getters }, { instanceAddress, packName, as }) {
-      const result = await downloadFrom(getters.authHost, instanceAddress, packName, as, getters.token)
+      const result = await downloadFrom(instanceAddress, packName, as, getters.authHost, getters.token)
 
       if (result.data === 'ok') {
         Message({
@@ -98,6 +125,25 @@ const packs = {
           duration: 5 * 1000
         })
       }
+    },
+    async FetchLocalEmojiPacks({ commit, getters, state }, page) {
+      const { data } = await listPacks(page, state.pageSize, getters.authHost, getters.token)
+      const { packs, count } = data
+      const updatedPacks = Object.keys(packs).reduce((acc, packName) => {
+        const { files, ...pack } = packs[packName]
+        acc[packName] = pack
+        return acc
+      }, {})
+      commit('SET_LOCAL_PACKS', updatedPacks)
+      commit('SET_LOCAL_PACKS_COUNT', count)
+      commit('SET_PAGE', page)
+    },
+    async FetchSinglePack({ getters, commit, state }, { name, page }) {
+      const { data } = await fetchPack(name, page, state.filesPageSize, getters.authHost, getters.token)
+      const { files, files_count } = data
+      commit('SET_PACK_FILES', { name, files })
+      commit('SET_FILES_COUNT', files_count)
+      commit('SET_FILES_PAGE', page)
     },
     async ImportFromFS({ getters }) {
       const result = await importFromFS(getters.authHost, getters.token)
@@ -136,12 +182,8 @@ const packs = {
         commit('UPDATE_LOCAL_PACK_PACK', { name: packName, pack: result.data })
       }
     },
-    SetActiveCollapseItems({ commit, state }, activeItems) {
-      commit('SET_ACTIVE_COLLAPSE_ITEMS', activeItems)
-    },
-    async SetLocalEmojiPacks({ commit, getters }) {
-      const { data } = await listPacks(getters.authHost)
-      commit('SET_LOCAL_PACKS', data)
+    SetActiveTab({ commit }, activeTab) {
+      commit('SET_ACTIVE_TAB', activeTab)
     },
     async SetRemoteEmojiPacks({ commit, getters }, { remoteInstance }) {
       const { data } = await listRemotePacks(getters.authHost, getters.token, remoteInstance)
@@ -152,10 +194,19 @@ const packs = {
     SetRemoteInstance({ commit }, instance) {
       commit('SET_REMOTE_INSTANCE', instance)
     },
-    async UpdateEmojiFile({ commit, getters }, { packName, shortcode, newShortcode, newFilename, force }) {
-      let result
+    async UpdateEmojiFile({ commit, dispatch, getters, state }, { packName, shortcode, newShortcode, newFilename, force }) {
+      const updatedPackFiles = Object.keys(state.localPacks[packName].files).reduce((acc, el) => {
+        if (el === shortcode) {
+          acc[newShortcode] = newFilename
+        } else {
+          acc[el] = state.localPacks[packName].files[el]
+        }
+        return acc
+      }, {})
+      commit('UPDATE_LOCAL_PACK_FILES', { name: packName, files: updatedPackFiles })
+
       try {
-        result = await updateEmojiFile(packName, shortcode, newShortcode, newFilename, force, getters.authHost, getters.token)
+        await updateEmojiFile(packName, shortcode, newShortcode, newFilename, force, getters.authHost, getters.token)
       } catch (_e) {
         return
       }
@@ -165,7 +216,7 @@ const packs = {
         duration: 5 * 1000
       })
 
-      commit('UPDATE_LOCAL_PACK_FILES', { name: packName, files: result.data })
+      dispatch('FetchSinglePack', { name: packName, page: state.currentFilesPage })
     },
     async UpdateLocalPackVal({ commit }, args) {
       commit('UPDATE_LOCAL_PACK_VAL', args)
@@ -173,4 +224,4 @@ const packs = {
   }
 }
 
-export default packs
+export default emojiPacks