+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - On Reports page add links to reported account and the author of the report 
 - In Notes add link to the note author's profile page 
 - In Moderation log add link to the actor's profile page
+- Support pagination of local emoji packs and files
 - Adds MRF Activity Expiration setting
 
 ### Changed